p4i integration Integrate 275107 from 4.3 to main: Fixes:    crash on certain 64 bit platforms RevBy:    Andreas 	AutoTest: qitemmodel et.al. Details:  pathconf() is documented to return a long, and 	  the value -1 when an undefined result is returned. 	  int(-1) and long(-1) differs when long is 8 bytes and 	  int only 4.
[git-p4: depot-paths = "//depot/qt/main/": change = 275108]

+++ src/corelib/io/qfsfileengine_iterator_unix.cpp
@@ -88,7 +88,7 @@ bool QFSFileEngineIterator::hasNext() const
         } else {
             // ### Race condition; we should use fpathconf and dirfd().
             long maxPathName = ::pathconf(QFile::encodeName(path()).data(), _PC_NAME_MAX);
-            if (maxPathName == -1)
+            if ((int) maxPathName == -1)
                 maxPathName = FILENAME_MAX;
             maxPathName += sizeof(dirent) + 1;
 #if defined(_POSIX_THREAD_SAFE_FUNCTIONS) && !defined(Q_OS_CYGWIN)
